pyproject.toml
[project]
name = "llm-study"
version = "0.1.0"
description = "LLM勉強会"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.50.1",
    "langchain>=1.0.5",
    "langchain-google-vertexai>=3.0.3",
    "python-dotenv>=1.2.1",
]
